parallel block statement
In the example above, fork-join block will be launched after executing the statement at 10 time units. Statements within this block will be executed in parallel and the first statement that will be launched will be the one where data is assigned a value of 8'h00 since the delay for that is 10 time units after the launch of the fork-join. After 10 more time units, the first statement will be launched and data will get the value 8'h11.