case statement syntax and example
The design module shown below has a 2-bit select signal to route one of the three other 3-bit inputs to the output signal called out . A case statement is used to assign the correct input to output based on the value of sel . Since sel is a 2-bit signal, it can have 22 combinations, 0 through 3. The default statement helps to set output to 0 if sel is 3.